% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerer.R
\name{source.bands}
\alias{source.bands}
\title{Visualize chromosome bands}
\usage{
source.bands(path.to.cbmrscripts)
}
\arguments{
\item{path.to.cbmrscripts}{Absolute path to cbmrscripts}
}
\value{
A series of functions
}
\description{
Visualize chromosome bands
}
