% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerer.R
\name{source.survscan}
\alias{source.survscan}
\title{Extensive survival analysis for a series of signatures}
\usage{
source.survscan(path.to.cbmrscripts,
  path.to.tcgadump = "/restricted/projectnb/montilab-p/personal/anthony/tcgadump",
  path.to.timer = "/restricted/projectnb/montilab-p/CBMrepositoryData/TCGA/tumorInfiltration/timer/TableS2.13059_2016_1028_MOESM3_ESM.txt")
}
\arguments{
\item{path.to.cbmrscripts}{Absolute path to cbmrscripts}

\item{path.to.tcgadump}{Absolute path to tcgadump}

\item{path.to.timer}{Absolute path to timer data}
}
\value{
A series of functions
}
\description{
Extensive survival analysis for a series of signatures
}
